Reorganizes card data structure and refactors game logic
Replaces scattered card definitions with a centralized data structure, improving maintainability and reducing redundancy.

Refactors game logic to streamline card management, movement, and interactions, enhancing modularity and readability.

Removes obsolete components and files while simplifying code handling for actions, effects, and summoning mechanisms.
