+++ src/components/ExtraMonsterZones.tsx
@@ -5,7 +5,6 @@ import { CARD_SIZE } from "@/const/card";
 
 interface ExtraMonsterZonesProps {
     extraMonsterZones: (CardInstance | null)[];
-    handleFieldCardClick: (card: CardInstance) => void;
     opponentField: {
         monsterZones: (CardInstance | null)[];
         spellTrapZones: (CardInstance | null)[];
@@ -13,11 +12,7 @@ interface ExtraMonsterZonesProps {
     };
 }
 
-export const ExtraMonsterZones: React.FC<ExtraMonsterZonesProps> = ({
-    extraMonsterZones,
-    handleFieldCardClick,
-    opponentField,
-}) => {
+export const ExtraMonsterZones: React.FC<ExtraMonsterZonesProps> = ({ extraMonsterZones, opponentField }) => {
     const cardSizeClass = CARD_SIZE.MEDIUM;
 
     return (
@@ -31,7 +26,6 @@ export const ExtraMonsterZones: React.FC<ExtraMonsterZonesProps> = ({
                 <FieldZone
                     card={extraMonsterZones[0]}
                     className={`${cardSizeClass} border-4 border-red-400`}
-                    onCardClick={handleFieldCardClick}
                     type={"extra_zone"}
                 />
 
@@ -41,18 +35,12 @@ export const ExtraMonsterZones: React.FC<ExtraMonsterZonesProps> = ({
                 <FieldZone
                     card={extraMonsterZones[1]}
                     className={`${cardSizeClass} border-4 border-red-400`}
-                    onCardClick={handleFieldCardClick}
                     type={"extra_zone"}
                 />
 
                 <div className={`${cardSizeClass}`}></div>
                 {/* 相手のフィールド魔法（右側） */}
-                <FieldZone
-                    type="field"
-                    card={opponentField?.fieldZone || null}
-                    className={CARD_SIZE.MEDIUM}
-                    onCardClick={handleFieldCardClick}
-                />
+                <FieldZone type="field" card={opponentField?.fieldZone || null} className={CARD_SIZE.MEDIUM} />
             </div>
         </div>
     );
