+++ src/components/FieldZone.tsx
@@ -13,6 +13,7 @@ interface FieldZoneProps {
     selected?: boolean;
     customSize?: string;
     reverse?: boolean;
+    disableActivate?: true;
 }
 
 export const FieldZone: React.FC<FieldZoneProps> = ({
@@ -25,6 +26,7 @@ export const FieldZone: React.FC<FieldZoneProps> = ({
     disabled,
     selected,
     customSize,
+    disableActivate,
 }) => {
     const handleClick = () => {
         if (card && onCardClick) {
@@ -48,11 +50,7 @@ export const FieldZone: React.FC<FieldZoneProps> = ({
                 onClick={handleClick}
             >
                 {card ? (
-                    <Card 
-                        card={card} 
-                        size="medium" 
-                        customSize={customSize} 
-                    />
+                    <Card card={card} size="medium" customSize={customSize} disableActivate={disableActivate} />
                 ) : type === "deck" ? (
                     <div className={`flex items-center justify-center ${textColor} text-xs w-full h-full`}>Deck</div>
                 ) : type === "extra_deck" ? (
