+++ src/components/GameBoard.tsx
@@ -1,15 +1,16 @@
-import React, { useEffect, useState, useCallback } from "react";
+import React, { useEffect, useState } from "react";
 import { useGameStore } from "@/store/gameStore";
 import { HandArea } from "./HandArea";
 import { PlayerField } from "./PlayerField";
 import { ExtraMonsterZones } from "./ExtraMonsterZones";
 import { ControlButtons } from "./ControlButtons";
-import { isMonsterCard, searchCombinationLinkSummon, searchCombinationXyzSummon } from "@/utils/gameUtils";
+import { searchCombinationLinkSummon, searchCombinationXyzSummon } from "@/utils/gameUtils";
 import type { CardInstance } from "@/types/card";
 import { HoveredCardDisplay } from "./HoveredCardDisplay";
 import { GraveyardModal } from "./GraveyardModal";
 import { ExtraDeckModal } from "./ExtraDeckModal";
 import { EffectQueueModal } from "./EffectQueueModal";
+import { isXyzMonster } from "@/utils/cardManagement";
 
 export const GameBoard: React.FC = () => {
     const gameState = useGameStore();
@@ -30,11 +31,11 @@ export const GameBoard: React.FC = () => {
         winner,
         isOpponentTurn,
         opponentField,
+        sendSpellToGraveyard,
     } = gameState;
 
     const [showGraveyard, setShowGraveyard] = useState(false);
     const [showExtraDeck, setShowExtraDeck] = useState(false);
-    const [selectedCard, setSelectedCard] = useState<string | null>(null);
 
     useEffect(() => {
         initializeGame();
@@ -46,41 +47,13 @@ export const GameBoard: React.FC = () => {
         // This would need to be added to gameStore
     };
 
-    const playCard = useCallback(
-        (_cardId: string) => {
-            // Implement card playing logic via gameStore
-            if (gameState.playCard) {
-                gameState.playCard(_cardId);
-            }
-        },
-        [gameState]
-    );
-
-    const setCard = (cardId: string | null) => {
-        setSelectedCard(cardId);
-    };
-
-    // Removed individual card effect handler - using generic handleEffect instead
-
     const startSpecialSummon = (monster: CardInstance, summonType: "link" | "xyz" | "synchro" | "fusion") => {
-        // Generic special summon handler
-        // Use the effect system for special summons
         if (summonType === "link" && gameState.startLinkSummon) {
             gameState.startLinkSummon(monster);
         } else if (summonType === "xyz" && gameState.startXyzSummon) {
             gameState.startXyzSummon(monster);
         }
     };
-
-    const sendToGraveyard = useCallback(
-        (cardInstance: CardInstance) => {
-            // Use gameStore's sendSpellToGraveyard
-            if (gameState.sendSpellToGraveyard) {
-                gameState.sendSpellToGraveyard(cardInstance);
-            }
-        },
-        [gameState]
-    );
 
     const checkExodiaWin = () => {
         // Implement exodia win check logic
@@ -97,12 +70,6 @@ export const GameBoard: React.FC = () => {
         // This would need to be added to gameStore
     };
 
-    const selectedCardInstance = hand.find((c) => c.id === selectedCard) || null;
-
-    const isLinkMonster = (card: CardInstance): boolean => {
-        return isMonsterCard(card.card) && (card.card as { monster_type?: string }).monster_type === "リンクモンスター";
-    };
-
     const canPerformLinkSummon = (linkMonster: CardInstance): boolean => {
         // アウローラドンの効果でリンク召喚制限がある場合
         if (gameState.isLinkSummonProhibited) {
@@ -116,77 +83,20 @@ export const GameBoard: React.FC = () => {
         );
     };
 
-    const isXyzMonster = (card: CardInstance): boolean => {
-        return (
-            isMonsterCard(card.card) && (card.card as { monster_type?: string }).monster_type === "エクシーズモンスター"
-        );
-    };
-
     const canPerformXyzSummon = (xyzMonster: CardInstance): boolean => {
-        if (!isXyzMonster(xyzMonster)) return false;
+        if (!isXyzMonster(xyzMonster.card)) return false;
 
         return searchCombinationXyzSummon(xyzMonster, gameState.field.extraMonsterZones, gameState.field.monsterZones);
     };
 
-    const handleEffect = (card: CardInstance) => {
-        // Generic effect handler - check card's effect conditions and push to queue
-        if (card.card.effect?.onIgnition) {
-            const canActivate = card.card.effect.onIgnition.condition(gameState, card);
-            if (canActivate) {
-                // Push effect activation to queue instead of direct execution
-                gameState.addEffectToQueue({
-                    id: card.id + "_ignition",
-                    type: "activate_spell",
-                    cardInstance: card,
-                    effectType: "card_effect_activation",
-                });
-            }
-        }
-    };
-
-    const handleFieldCardClick = (card: CardInstance) => {
-        // Generic field card click handler - push to queue instead of direct activation
-        if (card.card.effect?.onIgnition) {
-            const canActivate = card.card.effect.onIgnition.condition(gameState, card);
-            if (canActivate) {
-                // Push effect activation to queue instead of direct execution
-                gameState.addEffectToQueue({
-                    id: card.id + "_field_ignition",
-                    type: "activate_spell",
-                    cardInstance: card,
-                    effectType: "field_card_effect_activation",
-                });
-            }
-        }
-    };
-
     useEffect(() => {
         const currentEffect = effectQueue?.[0];
-        if (currentEffect?.type === "activate_spell") {
-            // Execute the card's effect when activate_spell is processed
-            if (
-                currentEffect.effectType === "card_effect_activation" ||
-                currentEffect.effectType === "field_card_effect_activation" ||
-                currentEffect.effectType === "spell_activation"
-            ) {
-                const card = currentEffect.cardInstance;
-                if (card.card.effect?.onIgnition) {
-                    // Execute the effect directly here since it's now properly queued
-                    card.card.effect.onIgnition.effect(gameState, card);
-                }
-                popQueue();
-            }
-        }
-        if (currentEffect?.type === "spell_end") {
-            sendToGraveyard(currentEffect.cardInstance);
-            popQueue();
-        }
         if (currentEffect?.type === "notify") {
             if (currentEffect.effectType === "judge") {
                 judgeWin();
             }
         }
-    }, [effectQueue, popQueue, sendToGraveyard, gameState]);
+    }, [effectQueue, popQueue, gameState, sendSpellToGraveyard]);
 
     useEffect(() => {
         checkExodiaWin();
@@ -217,7 +127,6 @@ export const GameBoard: React.FC = () => {
                                 {/* エクストラモンスターゾーン（相手と自分の間） */}
                                 <ExtraMonsterZones
                                     extraMonsterZones={field.extraMonsterZones}
-                                    handleFieldCardClick={handleFieldCardClick}
                                     opponentField={opponentField}
                                 />
 
@@ -227,10 +136,6 @@ export const GameBoard: React.FC = () => {
                                     deck={deck}
                                     extraDeck={extraDeck}
                                     graveyard={graveyard}
-                                    turn={turn}
-                                    phase={phase}
-                                    isOpponentTurn={isOpponentTurn}
-                                    handleFieldCardClick={handleFieldCardClick}
                                     setShowGraveyard={setShowGraveyard}
                                     setShowExtraDeck={setShowExtraDeck}
                                 />
@@ -238,28 +143,12 @@ export const GameBoard: React.FC = () => {
                         </div>
 
                         {/* 手札エリア */}
-                        <HandArea
-                            hand={hand}
-                            selectedCard={selectedCard}
-                            lifePoints={lifePoints}
-                            playCard={playCard}
-                            setCard={setCard}
-                            onCardHoverLeave={() => 1}
-                            handleEffect={handleEffect}
-                        />
+                        <HandArea hand={hand} lifePoints={lifePoints} />
                     </div>
                 </div>
 
                 {/* コントロールボタン */}
-                <ControlButtons
-                    selectedCardInstance={selectedCardInstance}
-                    selectedCard={selectedCard}
-                    isOpponentTurn={isOpponentTurn}
-                    nextPhase={nextPhase}
-                    playCard={playCard}
-                    setCard={setCard}
-                    initializeGame={initializeGame}
-                />
+                <ControlButtons isOpponentTurn={isOpponentTurn} nextPhase={nextPhase} initializeGame={initializeGame} />
 
                 {/* リンク */}
                 <div className="fixed bottom-4 right-8 space-y-2">
@@ -285,16 +174,13 @@ export const GameBoard: React.FC = () => {
                     onClose={() => setShowGraveyard(false)}
                     graveyard={graveyard}
                     gameState={gameState}
-                    handleEffect={handleEffect}
                 />
 
                 <ExtraDeckModal
                     isOpen={showExtraDeck}
                     onClose={() => setShowExtraDeck(false)}
                     extraDeck={extraDeck}
-                    isLinkMonster={isLinkMonster}
                     canPerformLinkSummon={canPerformLinkSummon}
-                    isXyzMonster={isXyzMonster}
                     canPerformXyzSummon={canPerformXyzSummon}
                     startLinkSummon={(monster) => startSpecialSummon(monster, "link")}
                     startXyzSummon={(monster) => startSpecialSummon(monster, "xyz")}
