+++ src/data/cards/monsters/extra.ts
@@ -1,24 +1,25 @@
-import type { ExtraMonster } from "../cards";
+import type { ExtraMonster } from "@/types/card";
 import {
     sumLevel,
     monsterFilter,
     hasLevelMonsterFilter,
     sumLink,
-    withUserConfirm,
-    withUserSelectCard,
-    sendCard,
-    withUserSummon,
     isMagicCard,
     isRitualMonster,
+    createCardInstance,
+} from "@/utils/cardManagement";
+import {
+    withUserConfirm,
+    withUserSelectCard,
+    withUserSummon,
     withTurnAtOneceCondition,
     withTurnAtOneceEffect,
-    createCardInstance,
-    summon,
     withOption,
-} from "../cards";
+} from "@/utils/effectUtils";
+import { sendCard, summon } from "@/utils/cardMovement";
 import type { CardInstance } from "@/types/card";
 import type { GameStore } from "@/store/gameStore";
-import { TOKEN } from "./token";
+import { TOKEN } from "../tokens";
 
 // Define extra monsters as literal objects with proper typing
 export const EXTRA_MONSTERS = [
@@ -38,7 +39,7 @@ export const EXTRA_MONSTERS = [
                 card.find((e) => monsterFilter(e.card) && !e.card.hasTuner)
             );
         },
-        text: "チューナー＋チューナー以外のモンスター１体以上\\n(1)：このカードがモンスターゾーンに存在する限り、お互いの手札・デッキから墓地へ送られるモンスターは墓地へは行かず除外される。\\n(2)：モンスターの効果・魔法・罠カードが発動した時、このカードをリリースして発動できる。その発動を無効にし破壊する。\\n(3)：このカードが墓地へ送られた場合に発動できる。デッキから儀式モンスター１体または儀式魔法カード１枚を手札に加える。",
+        text: "チューナー＋チューナー以外のモンスター１体以上\n(1)：このカードがモンスターゾーンに存在する限り、お互いの手札・デッキから墓地へ送られるモンスターは墓地へは行かず除外される。\n(2)：モンスターの効果・魔法・罠カードが発動した時、このカードをリリースして発動できる。その発動を無効にし破壊する。\n(3)：このカードが墓地へ送られた場合に発動できる。デッキから儀式モンスター１体または儀式魔法カード１枚を手札に加える。",
         image: "card100179270_1.jpg",
         hasDefense: true as const,
         hasLevel: true as const,
@@ -70,14 +71,16 @@ export const EXTRA_MONSTERS = [
                 },
             },
             onAnywhereToGraveyard: (gameState: GameStore, cardInstance: CardInstance) => {
-                const ritualCards = gameState.deck.filter((card) => {
-                    if (isRitualMonster(card.card)) return true;
-                    if (isMagicCard(card.card) && card.card.magic_type === "儀式魔法") return true;
-                    return false;
-                });
+                const ritualCards = (gameState: GameStore) =>
+                    gameState.deck.filter((card) => {
+                        if (isMagicCard(card.card) && card.card.magic_type === "儀式魔法") return true;
+                        if (isRitualMonster(card.card)) {
+                            return true;
+                        }
+                        return false;
+                    });
 
-                if (ritualCards.length === 0) return;
-
+                if (ritualCards(gameState).length === 0) return;
                 withUserConfirm(gameState, cardInstance, {}, (state, card) => {
                     withUserSelectCard(state, card, ritualCards, { select: "single" }, (state, _card, selected) => {
                         sendCard(state, selected[0], "Hand");
@@ -426,19 +429,24 @@ export const EXTRA_MONSTERS = [
             onIgnition: {
                 condition: (gameState: GameStore, cardInstance: CardInstance) => {
                     if (!withTurnAtOneceCondition(gameState, cardInstance, () => true)) return false;
-                    
-                    const faceUpMonsters = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].filter(
-                        (monster) => monster !== null && monster.position !== "back_defense" && monster.position !== "back"
+
+                    const faceUpMonsters = [
+                        ...gameState.field.monsterZones,
+                        ...gameState.field.extraMonsterZones,
+                    ].filter(
+                        (monster) =>
+                            monster !== null && monster.position !== "back_defense" && monster.position !== "back"
                     );
-                    
+
                     return faceUpMonsters.length > 0;
                 },
                 effect: (gameState: GameStore, cardInstance: CardInstance) => {
                     withTurnAtOneceEffect(gameState, cardInstance, (state, card) => {
                         const faceUpMonsters = [...state.field.monsterZones, ...state.field.extraMonsterZones].filter(
-                            (monster) => monster !== null && monster.position !== "back_defense" && monster.position !== "back"
+                            (monster) =>
+                                monster !== null && monster.position !== "back_defense" && monster.position !== "back"
                         );
-                        
+
                         withUserSelectCard(
                             state,
                             card,
@@ -487,19 +495,24 @@ export const EXTRA_MONSTERS = [
             onIgnition: {
                 condition: (gameState: GameStore, cardInstance: CardInstance) => {
                     if (!withTurnAtOneceCondition(gameState, cardInstance, () => true)) return false;
-                    
-                    const faceUpMonsters = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].filter(
-                        (monster) => monster !== null && monster.position !== "back_defense" && monster.position !== "back"
+
+                    const faceUpMonsters = [
+                        ...gameState.field.monsterZones,
+                        ...gameState.field.extraMonsterZones,
+                    ].filter(
+                        (monster) =>
+                            monster !== null && monster.position !== "back_defense" && monster.position !== "back"
                     );
-                    
+
                     return faceUpMonsters.length > 0;
                 },
                 effect: (gameState: GameStore, cardInstance: CardInstance) => {
                     withTurnAtOneceEffect(gameState, cardInstance, (state, card) => {
                         const faceUpMonsters = [...state.field.monsterZones, ...state.field.extraMonsterZones].filter(
-                            (monster) => monster !== null && monster.position !== "back_defense" && monster.position !== "back"
+                            (monster) =>
+                                monster !== null && monster.position !== "back_defense" && monster.position !== "back"
                         );
-                        
+
                         withUserSelectCard(
                             state,
                             card,
@@ -540,10 +553,11 @@ export const EXTRA_MONSTERS = [
                 },
                 effect: (gameState: GameStore, cardInstance: CardInstance) => {
                     // Release self to protect another monster
-                    const targetMonsters = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].filter(
-                        (monster): monster is CardInstance => monster !== null && monster.id !== cardInstance.id
-                    );
-                    
+                    const targetMonsters = [
+                        ...gameState.field.monsterZones,
+                        ...gameState.field.extraMonsterZones,
+                    ].filter((monster): monster is CardInstance => monster !== null && monster.id !== cardInstance.id);
+
                     if (targetMonsters.length > 0) {
                         withUserSelectCard(
                             gameState,
@@ -560,10 +574,11 @@ export const EXTRA_MONSTERS = [
             },
             onAnywhereToGraveyard: (gameState: GameStore, cardInstance: CardInstance) => {
                 // Special summon from graveyard when normal summoned monster is destroyed
-                const normalSummonedMonsters = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].filter(
-                    (monster) => monster !== null && monster.summonedBy === "Normal"
-                );
-                
+                const normalSummonedMonsters = [
+                    ...gameState.field.monsterZones,
+                    ...gameState.field.extraMonsterZones,
+                ].filter((monster) => monster !== null && monster.summonedBy === "Normal");
+
                 if (normalSummonedMonsters.length > 0) {
                     withUserConfirm(gameState, cardInstance, {}, (state, card) => {
                         withUserSummon(state, card, card, () => {});
