+++ src/utils/summonUtils.ts
@@ -1,51 +1,18 @@
-import type { Card, CardInstance, MonsterCard } from "@/types/card";
+import type { Card, CardInstance } from "@/types/card";
 import type { GameState } from "@/types/game";
-import { isMonsterCard } from "./gameUtils";
-import type { GameStore } from "@/store/gameStore";
+import { isMagicCard, isTrapCard, monsterFilter } from "./cardManagement";
 
 export const canNormalSummon = (gameState: GameState, card: CardInstance): boolean => {
-    if (!isMonsterCard(card.card)) return false;
-    if (card.card.card_name === "神聖なる魂") return false;
-
+    if (!monsterFilter(card.card)) return false;
     if (gameState.hasNormalSummoned) return false;
     if (gameState.phase !== "main1" && gameState.phase !== "main2") return false;
 
-    const monsterCard = card.card as MonsterCard;
-
     // 特殊召喚モンスターは通常召喚できない
-    if (card.card.card_type === "特殊召喚・効果モンスター") return false;
+    if (!card.card.canNormalSummon) return false;
 
-    // 儀式モンスターは通常召喚できない
-    if (card.card.card_type === "儀式・効果モンスター") return false;
-
-    // 融合・シンクロ・エクシーズ・リンクモンスターは通常召喚できない
-    if (
-        card.card.card_type === "融合モンスター" ||
-        card.card.card_type === "シンクロモンスター" ||
-        card.card.card_type === "エクシーズモンスター" ||
-        card.card.card_type === "リンクモンスター"
-    )
-        return false;
-
-    // レベル5以上はリリースが必要
-    if (monsterCard.level && monsterCard.level >= 5) {
-        const requiredTributes = monsterCard.level >= 7 ? 2 : 1;
-        const availableTributes = gameState.field.monsterZones.filter((m) => m !== null).length;
-        return availableTributes >= requiredTributes;
-    }
     const hasEmpty = gameState.field.monsterZones.filter((e) => e === null).length > 0;
 
     return hasEmpty;
-};
-
-export const getRequiredTributes = (card: Card): number => {
-    if (!isMonsterCard(card)) return 0;
-    const monsterCard = card as MonsterCard;
-
-    if (!monsterCard.level) return 0;
-    if (monsterCard.level >= 7) return 2;
-    if (monsterCard.level >= 5) return 1;
-    return 0;
 };
 
 export const findEmptyMonsterZone = (gameState: GameState): number => {
@@ -67,382 +34,18 @@ export const findEmptySpellTrapZone = (gameState: GameState): number => {
     return -1;
 };
 
-export const canActivateSacredSoul = (gameState: GameStore) => {
-    return (
-        gameState.hand.find((e) => e.card.card_name === "神聖なる魂") &&
-        gameState.graveyard.filter((e) => {
-            if (!isMonsterCard(e.card)) {
-                return false;
-            }
-            const typed = e.card as MonsterCard;
-            return typed.attribute === "光属性";
-        }).length >= 2 &&
-        gameState.field.monsterZones.filter((e) => e === null).length > 0
-    );
-};
-
-export const canActivateSpell = (gameState: GameState, card: Card): boolean => {
-    const spellTypes = ["通常魔法", "速攻魔法", "永続魔法", "フィールド魔法", "装備魔法", "儀式魔法"];
-
-    if (!spellTypes.includes(card.card_type)) return false;
-
-    // メインフェイズのみ発動可能（速攻魔法は除く）
-    if (card.card_type !== "速攻魔法" && gameState.phase !== "main1" && gameState.phase !== "main2") {
-        return false;
-    }
-
-    // ジャック・イン・ザ・ハンドの特別な発動条件チェック
-    if (card.card_name === "ジャック・イン・ザ・ハンド") {
-        return canActivateJackInHand(gameState);
-    }
-
-    // おろかな埋葬の特別な発動条件チェック
-    if (card.card_name === "おろかな埋葬") {
-        return canActivateFoolishBurial(gameState);
-    }
-
-    // 金満で謙虚な壺の特別な発動条件チェック
-    if (card.card_name === "金満で謙虚な壺") {
-        return canActivateExtravagance(gameState);
-    }
-
-    // ワン・フォー・ワンの特別な発動条件チェック
-    if (card.card_name === "ワン・フォー・ワン") {
-        return canActivateOneForOne(gameState);
-    }
-
-    // エマージェンシー・サイバーの特別な発動条件チェック
-    if (card.card_name === "エマージェンシー・サイバー") {
-        return canActivateEmergencyCyber(gameState);
-    }
-
-    // 現在、金満で謙虚な壺による制限対象のカードはなし
-    // （すべて「手札に加える」効果であり、「ドロー」効果ではないため）
-
-    // 高等儀式術の特別な発動条件チェック
-    if (card.card_name === "高等儀式術") {
-        return canActivateAdvancedRitual(gameState);
-    }
-
-    // 流星輝巧群の特別な発動条件チェック
-    if (card.card_name === "流星輝巧群") {
-        return canActivateMeteorKikougun(gameState);
-    }
-
-    // 儀式の準備の特別な発動条件チェック
-    if (card.card_name === "儀式の準備") {
-        return canActivatePreparationForRitual(gameState);
-    }
-
-    // 竜輝巧－ファフニールの特別な発動条件チェック
-    if (card.card_name === "竜輝巧－ファフニール") {
-        return canActivateFafnir(gameState);
-    }
-
-    // 極超の竜輝巧の特別な発動条件チェック
-    if (card.card_name === "極超の竜輝巧") {
-        return canActivateDreitronNova(gameState);
-    }
-
-    // フィールド魔法は専用ゾーンへ（盆回し制限をチェック）
-    if (card.card_type === "フィールド魔法") {
-        if (gameState.bonmawashiRestriction) {
-            return false;
-        }
-        return true;
-    }
-
-    // OCGルール準拠: 通常・速攻・儀式魔法も魔法・罠ゾーンに空きが必要
-    if (card.card_type === "通常魔法" || card.card_type === "速攻魔法" || card.card_type === "儀式魔法") {
-        return findEmptySpellTrapZone(gameState) !== -1;
-    }
-
-    // 永続魔法・装備魔法も魔法・罠ゾーンに空きが必要
-    return findEmptySpellTrapZone(gameState) !== -1;
-};
-
-export const canActivateJackInHand = (gameState: GameState): boolean => {
-    // デッキから異なるレベル1モンスター3体を探す
-    if (gameState.hasActivatedJackInTheHand) {
-        return false; // このターンに既に発動済み
-    }
-    const level1Monsters = gameState.deck.filter((c) => {
-        if (!isMonsterCard(c.card)) return false;
-        const monster = c.card as { level?: number };
-        return monster.level === 1;
-    });
-
-    // 異なるカード名のモンスターを数える
-    const uniqueNames = new Set<string>();
-    for (const monster of level1Monsters) {
-        uniqueNames.add(monster.card.card_name);
-    }
-
-    // 異なるカード名のレベル1モンスターが3体以上必要
-    return uniqueNames.size >= 3;
-};
-
-export const canActivateFoolishBurial = (gameState: GameState): boolean => {
-    // デッキにモンスターが1枚以上必要
-    const monstersInDeck = gameState.deck.filter((c) => isMonsterCard(c.card));
-    return monstersInDeck.length >= 1;
-};
-
-export const canActivateExtravagance = (gameState: GameState): boolean => {
-    if (gameState.hasActivatedExtravagance) {
-        return false;
-    } // 金満で謙虚な壺を発動したフラグを立てる
-
-    // EXデッキが3枚以上必要
-    if (gameState.extraDeck.length < 3) return false;
-
-    // このターンにカードの効果でドローしていない
-    if (gameState.hasDrawnByEffect) return false;
-
-    return true;
-};
-
-export const canActivateOneForOne = (gameState: GameState): boolean => {
-    // 手札にモンスターが1体以上必要
-    const handMonsters = gameState.hand.filter((c) => isMonsterCard(c.card));
-    if (handMonsters.length === 0) return false;
-
-    // デッキにレベル1モンスターが1体以上必要
-    const level1Monsters = gameState.deck.filter((c) => {
-        if (!isMonsterCard(c.card)) return false;
-        const monster = c.card as { level?: number };
-        return monster.level === 1;
-    });
-    const hasEmpty = gameState.field.monsterZones.filter((e) => e === null).length > 0;
-
-    return level1Monsters.length > 0 && hasEmpty;
-};
-
-export const canActivateDreitronNova = (gameState: GameState): boolean => {
-    if (gameState.hasActivatedDreitronNova === true) {
-        return false;
-    }
-    const monsters = gameState.deck.filter((c) => {
-        if (!isMonsterCard(c.card)) return false;
-        return c.card.card_name.includes("竜輝巧");
-    });
-    const hasEmpty = gameState.field.monsterZones.filter((e) => e === null).length > 0;
-
-    return monsters.length > 0 && hasEmpty;
-};
-
-export const canActivateEmergencyCyber = (gameState: GameState): boolean => {
-    // このターンに既に発動済みの場合は発動不可
-    if (gameState.hasActivatedEmergencyCyber) {
-        return false;
-    }
-
-    // デッキに対象となるモンスターが1体以上必要
-    const targetMonsters = gameState.deck.filter((c) => {
-        if (!isMonsterCard(c.card)) return false;
-        const monster = c.card as {
-            card_name?: string;
-            race?: string;
-            attribute?: string;
-            card_type?: string;
-        };
-
-        // サイバー・ドラゴンモンスター
-        if (monster.card_name?.includes("サイバー・ドラゴン")) return true;
-
-        // 通常召喚できない機械族・光属性モンスター
-        if (
-            monster.race === "機械族" &&
-            monster.attribute === "光属性" &&
-            monster.card_type === "特殊召喚・効果モンスター"
-        ) {
-            return true;
-        }
-
-        return false;
-    });
-
-    return targetMonsters.length > 0;
-};
-
-export const canActivateAdvancedRitual = (gameState: GameState): boolean => {
-    // 手札に儀式モンスターが1体以上必要
-    const ritualMonsters = gameState.hand.filter(
-        (c) => isMonsterCard(c.card) && c.card.card_type === "儀式・効果モンスター"
-    );
-
-    if (ritualMonsters.length === 0) {
-        return false;
-    }
-
-    // 手札の儀式モンスターの最小レベルを取得
-    const minRitualLevel = Math.min(
-        ...ritualMonsters.map((c) => {
-            const monster = c.card as { level?: number };
-            return monster.level || 0;
-        })
-    );
-
-    // デッキの通常モンスターのレベル合計を計算
-    const normalMonsters = gameState.deck.filter((c) => isMonsterCard(c.card) && c.card.card_type === "通常モンスター");
-
-    const totalNormalLevel = normalMonsters.reduce((sum, c) => {
-        const monster = c.card as { level?: number };
-        return sum + (monster.level || 0);
-    }, 0);
-
-    // デッキの通常モンスターレベル合計が手札の儀式モンスター最小レベル以上必要
-    const canActivate = totalNormalLevel >= minRitualLevel;
-    const hasEmpty = gameState.field.monsterZones.filter((e) => e === null).length > 0;
-
-    return canActivate && hasEmpty;
-};
-
-export const canActivateHokyuYoin = (gameState: GameState): boolean => {
-    // 墓地にモンスターが5体以上必要
-    const graveyardMonsters = gameState.graveyard.filter((c) => isMonsterCard(c.card));
-    if (graveyardMonsters.length < 5) return false;
-
-    // 効果モンスター以外の攻撃力1500以下のモンスターが1体以上必要
-    const targetMonsters = gameState.graveyard.filter((c) => {
-        if (!isMonsterCard(c.card)) return false;
-        const monster = c.card as { card_type?: string; attack?: number };
-        return monster.card_type === "通常モンスター" && (monster.attack || 9999) <= 1500;
-    });
-
-    return targetMonsters.length >= 1;
-};
-
-export const canActivateFafnir = (gameState: GameState): boolean => {
-    // このターンに既に発動済みの場合は発動不可
-    if (gameState.hasActivatedFafnir) {
-        return false;
-    }
-
-    return true;
-};
-
-export const canActivatePreparationForRitual = (gameState: GameState): boolean => {
-    const deckRitualMonsters = gameState.deck.filter(
-        (c) => isMonsterCard(c.card) && c.card.card_type === "儀式・効果モンスター"
-    );
-
-    if (deckRitualMonsters.length === 0) {
-        return false;
-    }
-    return true;
-};
-
-export const canActivateMeteorKikougun = (gameState: GameState): boolean => {
-    console.log("Checking Meteor Kikougun activation conditions...");
-
-    // 手札・墓地に儀式モンスターが1体以上必要
-    const handRitualMonsters = gameState.hand.filter(
-        (c) => isMonsterCard(c.card) && c.card.card_type === "儀式・効果モンスター"
-    );
-
-    const graveyardRitualMonsters = gameState.graveyard.filter(
-        (c) => isMonsterCard(c.card) && c.card.card_type === "儀式・効果モンスター"
-    );
-
-    if (handRitualMonsters.length === 0 && graveyardRitualMonsters.length === 0) {
-        return false;
-    }
-    const allRitualMonsters = [...handRitualMonsters, ...graveyardRitualMonsters];
-    const ritualMonstersMinAttack = Math.min(...allRitualMonsters.map((c) => (c.card as { attack: number }).attack));
-
-    // 手札・フィールドに機械族モンスターが1体以上必要
-    const handMachineMonsters = gameState.hand.filter((c) => {
-        if (!isMonsterCard(c.card)) return false;
-        const monster = c.card as { race?: string };
-        return monster.race === "機械族";
-    });
-
-    const fieldMachineMonsters = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].filter(
-        (c): c is CardInstance => {
-            if (!c || !isMonsterCard(c.card)) return false;
-            const monster = c.card as { race?: string };
-            return monster.race === "機械族";
-        }
-    );
-
-    const sumPower =
-        handMachineMonsters.reduce((sum, c) => sum + (c.card as { attack: number }).attack, 0) +
-        fieldMachineMonsters.reduce((sum, c) => sum + (c.card as { attack: number }).attack, 0);
-
-    const canActivate =
-        handMachineMonsters.length + fieldMachineMonsters.length >= 1 && sumPower >= ritualMonstersMinAttack;
-
-    const hasEmpty =
-        gameState.field.monsterZones.filter((e) => e === null).length > 0 ||
-        (gameState.field.monsterZones.filter((e) => e === null).length === 0 ? fieldMachineMonsters.length > 0 : false);
-
-    return canActivate && hasEmpty;
-};
-
-export const canActivateBanAlpha = (gameState: GameStore): boolean => {
-    if (gameState.hasActivatedBanAlpha) {
-        return false;
-    }
-    return canActivateDreitrons(gameState, "竜輝巧－バンα");
-};
-
-export const canActivateDreitrons = (gameState: GameStore, dreitronName: string) => {
-    const handTargets = gameState.hand.filter((c) => {
-        if (!isMonsterCard(c.card)) return false;
-        const isDrytron =
-            (c.card.card_name.includes("竜輝巧") || c.card.card_name.includes("ドライトロン")) &&
-            c.card.card_name !== dreitronName;
-        const isRitual = c.card.card_type === "儀式・効果モンスター";
-        return isDrytron || isRitual;
-    });
-
-    const fieldTargets = gameState.field.monsterZones.filter((c): c is CardInstance => {
-        if (!c || !isMonsterCard(c.card)) return false;
-        const isDrytron =
-            (c.card.card_name.includes("竜輝巧") || c.card.card_name.includes("ドライトロン")) &&
-            c.card.card_name !== dreitronName;
-        const isRitual = c.card.card_type === "儀式・効果モンスター";
-        return isDrytron || isRitual;
-    });
-
-    const hasEmpty =
-        gameState.field.monsterZones.filter((e) => e === null).length > 0 ||
-        (gameState.field.monsterZones.filter((e) => e === null).length === 0 && fieldTargets.length > 0);
-    return handTargets.length + fieldTargets.length >= 1 && hasEmpty;
-};
-
-export const canActivateAruZeta = (gameState: GameStore): boolean => {
-    // このターンに既に発動済みの場合は発動不可
-    if (gameState.hasActivatedAruZeta) {
-        return false;
-    }
-    return canActivateDreitrons(gameState, "竜輝巧－アルζ");
-};
-
-export const canActivateEruGanma = (gameState: GameStore): boolean => {
-    if (gameState.hasActivatedEruGanma) {
-        return false;
-    }
-    return canActivateDreitrons(gameState, "竜輝巧－エルγ");
-};
+const settableTrapTypes = ["通常罠", "永続罠", "カウンター罠"];
 
 export const canSetSpellTrap = (gameState: GameState, card: Card): boolean => {
-    const settableTypes = [
-        "通常魔法",
-        "速攻魔法",
-        "永続魔法",
-        "装備魔法",
-        "儀式魔法",
-        "通常罠カード",
-        "永続罠カード",
-        "カウンター罠カード",
-    ];
-
-    if (!settableTypes.includes(card.card_type)) return false;
-
-    // フィールド魔法はセットできない（直接発動のみ）
-    if (card.card_type === "フィールド魔法") return false;
+    if (!isMagicCard(card) && !isTrapCard(card)) {
+        return false;
+    }
+    if (isMagicCard(card) && card.magic_type === "フィールド魔法") {
+        return false;
+    }
+    if (isTrapCard(card) && !settableTrapTypes.includes(card.trap_type)) {
+        return false;
+    }
 
     // メインフェイズのみセット可能
     if (gameState.phase !== "main1" && gameState.phase !== "main2") {
@@ -451,130 +54,4 @@ export const canSetSpellTrap = (gameState: GameState, card: Card): boolean => {
 
     // 魔法・罠ゾーンに空きが必要
     return findEmptySpellTrapZone(gameState) !== -1;
-};
-
-// 後方互換性のため残しておく
-export const canSetTrap = canSetSpellTrap;
-
-export const canActivateBeatrice = (gameState: GameState): boolean => {
-    // 1ターンに1度の制限チェック
-    if (gameState.hasActivatedBeatriceEffect) {
-        return false;
-    }
-
-    // フィールドに永遠の淑女 ベアトリーチェがいるかチェック
-    const beatriceOnField = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].find(
-        (monster) => monster && monster.card.card_name === "永遠の淑女 ベアトリーチェ"
-    );
-
-    if (!beatriceOnField) {
-        return false;
-    }
-
-    // X素材があるかチェック
-    if (!beatriceOnField.materials || beatriceOnField.materials.length === 0) {
-        return false;
-    }
-
-    // デッキにカードがあるかチェック
-    if (gameState.deck.length === 0) {
-        return false;
-    }
-
-    return true;
-};
-
-export const canActivatePtolemyM7 = (gameState: GameState): boolean => {
-    if (gameState.hasActivatedPtolemyM7Effect) {
-        return false;
-    }
-
-    const ptolemyOnField = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].find(
-        (monster) => monster && monster.card.card_name === "セイクリッド・トレミスM7"
-    );
-
-    if (!ptolemyOnField || !ptolemyOnField.materials || ptolemyOnField.materials.length === 0) {
-        return false;
-    }
-
-    // フィールド・墓地にモンスターがあるかチェック
-    const fieldMonsters = [
-        ...gameState.field.monsterZones,
-        ...gameState.field.extraMonsterZones,
-        ...gameState.opponentField.monsterZones,
-    ].filter((monster) => monster !== null);
-
-    const graveyardMonsters = gameState.graveyard.filter((card) => card.card.card_type?.includes("モンスター"));
-
-    return fieldMonsters.length > 0 || graveyardMonsters.length > 0;
-};
-
-export const canActivateAuroradon = (gameState: GameState): boolean => {
-    if (gameState.hasActivatedAuroradonEffect) {
-        return false;
-    }
-
-    const auroradonOnField = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].find(
-        (monster) => monster && monster.card.card_name === "幻獣機アウローラドン"
-    );
-
-    if (!auroradonOnField) {
-        return false;
-    }
-
-    // 他のモンスターがフィールドにいるかチェック
-    const otherMonsters = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].filter(
-        (monster) => monster !== null
-    );
-
-    if (otherMonsters.length === 0) {
-        return false;
-    }
-
-    return true;
-};
-
-export const canActivateUnionCarrier = (gameState: GameState): boolean => {
-    if (gameState.hasActivatedUnionCarrierEffect) {
-        return false;
-    }
-
-    const unionCarrierOnField = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].find(
-        (monster) => monster && monster.card.card_name === "ユニオン・キャリアー"
-    );
-
-    if (!unionCarrierOnField) {
-        return false;
-    }
-
-    // フィールドに表側表示モンスターがあるかチェック
-    const faceUpMonsters = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].filter(
-        (monster) => monster !== null && monster.position !== "facedown" && monster.position !== "facedown_defense"
-    );
-    const hasEmpty = gameState.field.spellTrapZones.filter((e) => e === null).length > 0;
-
-    return faceUpMonsters.length > 0 && hasEmpty;
-};
-
-export const canActivateMeteorKikougunGraveyard = (gameState: GameState): boolean => {
-    if (gameState.hasActivatedMeteorKikougunGraveyardEffect) {
-        return false;
-    }
-
-    // 墓地に流星輝巧群があるかチェック
-    const meteorInGraveyard = gameState.graveyard.find((card) => card.card.card_name === "流星輝巧群");
-
-    if (!meteorInGraveyard) {
-        return false;
-    }
-
-    // フィールドにドライトロンモンスターがあるかチェック
-    const drytronMonstersOnField = [...gameState.field.monsterZones, ...gameState.field.extraMonsterZones].filter(
-        (monster) => {
-            if (!monster) return false;
-            return monster.card.card_name.includes("竜輝巧") || monster.card.card_name.includes("ドライトロン");
-        }
-    );
-
-    return drytronMonstersOnField.length > 0;
 };
